+++ src/index.tsx
@@ -1,30 +1,57 @@
 // App.tsx
+import { client } from '@api/network/client';
+import { useAppSelector } from '@hooks/redux';
 import { NavigationContainer } from '@react-navigation/native';
 import { createStackNavigator } from '@react-navigation/stack';
-import React, { useEffect, useState } from 'react';
+import { getToken, isAuth } from '@store/reducers/authSlice';
+import { setupStore } from '@store/store';
+import React, { useEffect } from 'react';
+import { Text } from 'react-native';
+import { Provider } from 'react-redux';
+import { PersistGate } from 'redux-persist/integration/react';
 
 import LoginScreen from './screens/auth/LoginScreen';
 import SignupScreen from './screens/auth/SignupScreen';
 import VerificationScreen from './screens/auth/VerificationScreen';
 import FeedScreen from './screens/feed/feedScreen';
 
-const Stack = createStackNavigator();
+const { store, persistor } = setupStore();
 
-const App: React.FC = () => {
-	const [isLoggedIn, setIsLoggedIn] = useState(false);
+const Stack = createStackNavigator();
+/*
+
+// TODO: To initialize the store, you need to call :
+const { store, persistor } = setupStore();
+<Provider store={store}>
+	<PersistGate loading={<Text>Loading...</Text>} persistor={persistor}>
+		<View></View>
+	</PersistGate>
+</Provider>
+
+// TODO: To retrieve the data :
+import { useAppSelector } from '@hooks/redux';
+import { retrieveUsername } from '@store/reducers/userSlice';
+const username = useAppSelector(retrieveUsername);
+
+// TODO: To dispatch the data :
+import { useAppDispatch } from '@hooks/redux';
+import { setUsername } from '@store/reducers/userSlice';
+const dispatch = useAppDispatch();
+dispatch(serUsername("bob"));
+*/
+
+const App = () => {
+	const token = useAppSelector(getToken);
+	const isLoggedIn = useAppSelector(isAuth);
 
 	useEffect(() => {
-		// const checkLoginStatus = async () => {
-		// 	try {
-		// 		const isLoggedInValue = await AsyncStorage.getItem('token');
-		// 		setIsLoggedIn(Boolean(isLoggedInValue));
-		// 	} catch (error) {
-		// 		console.error("Erreur lors de la récupération de l'état de connexion : ", error);
-		// 	}
-		// };
-		setIsLoggedIn(false);
-		//checkLoginStatus();
-	}, []);
+		const checkLoginStatus = async () => {
+			if (token) {
+				client.defaults.headers.post.Authorization = `Bearer ${token}`;
+			}
+		};
+		checkLoginStatus();
+	});
 
 	return (
 		<NavigationContainer>
@@ -52,4 +79,14 @@ const App: React.FC = () => {
 	);
 };
 
-export default App;
+const AppProvider = () => {
+	return (
+		<Provider store={store}>
+			<PersistGate loading={<Text>Loading ....</Text>} persistor={persistor}>
+				<App />
+			</PersistGate>
+		</Provider>
+	);
+};
+
+export default AppProvider;