+++ src/screens/auth/LoginScreen.tsx
@@ -1,4 +1,7 @@
-import AsyncStorage from '@react-native-async-storage/async-storage';
+import { loginUser } from '@api/auth.req';
+import { client } from '@api/network/client';
+import { useAppDispatch } from '@hooks/redux';
+import { setAuthentication, setToken } from '@store/reducers/authSlice';
 import React, { useState } from 'react';
 import { Image, StyleSheet, Text, TextInput, TouchableOpacity, View } from 'react-native';
 
@@ -13,14 +16,19 @@ export type LoginScreen = {
 const LoginScreen: LoginScreen = ({ navigation }) => {
 	const [email, setEmail] = useState('');
 	const [password, setPassword] = useState('');
+	const dispatch = useAppDispatch();
 
 	const handleLogin = async () => {
 		try {
-			//connexion a l'api
-			await AsyncStorage.setItem('token', 'le token recu');
-			navigation.navigate('Feed');
+			const res = await loginUser({ email, password });
+			if (res.status === 200) {
+				dispatch(setToken(res.data));
+				dispatch(setAuthentication(true));
+				client.defaults.headers.post.Authorization = `Bearer ${res.data}`;
+				navigation.navigate('Feed');
+			}
 		} catch (error) {
-			console.error('Erreur de connexion : ', error);
+			console.log(error);
 		}
 	};