+++ CHANGELOG.md
@@ -1,20 +1,71 @@
-# 1.0.0-dev.1 (2024-03-06)
-
-
-### Bug Fixes
-
-* **babel:** The babel config prevent ios and android to build because it was missconfigured ([4961996](https://github.com/Maghwyn/meme-rn-app/commit/4961996fe04a73b6740b3402c88f4edb98ace7e5))
-* **babel:** The babel config prevent ios and android to build because it was missconfigured ([8c3b5c6](https://github.com/Maghwyn/meme-rn-app/commit/8c3b5c6784ead53311e3f73dea37e8dcae990ffa))
-* **ci:** Ignore the scripts when installing with yarn ([f79e783](https://github.com/Maghwyn/meme-rn-app/commit/f79e783d1f835f3253e7fbc6aa3997b563853dbc))
-* **prettier:** Prettier failed to work because of missing packages + lint ([99bf4fa](https://github.com/Maghwyn/meme-rn-app/commit/99bf4fa97be7ef6daa966a188cab1faafab30a8a))
-* **prettier:** Prettier failed to work because of missing packages + lint ([a2fc4bf](https://github.com/Maghwyn/meme-rn-app/commit/a2fc4bf497bcd106b945d58ecf347bf3ceeeaac6))
-* **yarn:** Update yarn lock file ([31da3ec](https://github.com/Maghwyn/meme-rn-app/commit/31da3ec72a98a39dbb6a851ace9d0ccfc52144eb))
-* **yarn:** Update yarn lock file ([587f896](https://github.com/Maghwyn/meme-rn-app/commit/587f896caee168437fb66a8eb8682a822bf0edcf))
-* **yarn:** Update yarn lock file ([4aaf26c](https://github.com/Maghwyn/meme-rn-app/commit/4aaf26c9499a6c0c0acbea5c27a8c9d2b257ac66))
+# [1.0.0-dev.6](https://github.com/Maghwyn/meme-rn-app/compare/v1.0.0-dev.5...v1.0.0-dev.6) (2024-03-07)
 
 
 ### Features
 
-* **cd:** Implement the basic cd before deloyment. Wip ci ([9dfd2c4](https://github.com/Maghwyn/meme-rn-app/commit/9dfd2c439a2c21529ea25be81541ec71517dbc99))
-* **commitling:** Install commitlint for semantic-release support ([5299df9](https://github.com/Maghwyn/meme-rn-app/commit/5299df9e3c038f81736bbcd7edb3b6597e2b53af))
-* **commitling:** Install commitlint for semantic-release support ([40b164c](https://github.com/Maghwyn/meme-rn-app/commit/40b164c685132f5b909c338d2455df8dfef95d6a))
+* **api:** Create the api requests for the memes and the upload ([6543cce](https://github.com/Maghwyn/meme-rn-app/commit/6543ccebc4c1e12e60da8a47bca96c55ed677433))
+* **store:** Implement the meme store and fix the duplicated import in the other store ([b5d2fe6](https://github.com/Maghwyn/meme-rn-app/commit/b5d2fe6416ee3ee64718c676802b7f484e2497ea))
+* **utils:** Define some string utilities ([d86f63c](https://github.com/Maghwyn/meme-rn-app/commit/d86f63cab495ec8ff67791780e75f21be2b15e2e))
+
+# [1.0.0-dev.5](https://github.com/Maghwyn/meme-rn-app/compare/v1.0.0-dev.4...v1.0.0-dev.5) (2024-03-07)
+
+
+### Bug Fixes
+
+* **eslint:** Restriction rule for max lignes for a function at 100 lines ([75b28c8](https://github.com/Maghwyn/meme-rn-app/commit/75b28c8f49df7aa5a0b32d2396e4d2e14e326337))
+
+
+### Features
+
+* **auth:** Dispatch jwt token & state boolean for user isAuth ([68e2792](https://github.com/Maghwyn/meme-rn-app/commit/68e27928abb6f1ca976004fa1b4d06b3835b1b85))
+
+# [1.0.0-dev.4](https://github.com/Maghwyn/meme-rn-app/compare/v1.0.0-dev.3...v1.0.0-dev.4) (2024-03-07)
+
+
+### Features
+
+* **feed-view:** like and comment button ([11a038f](https://github.com/Maghwyn/meme-rn-app/commit/11a038f6c12f6a3e0c3df1be7bfad49770231883))
+
+# [1.0.0-dev.3](https://github.com/Maghwyn/meme-rn-app/compare/v1.0.0-dev.2...v1.0.0-dev.3) (2024-03-07)
+
+
+### Bug Fixes
+
+* **auth:** call http client in signup & login & verification code screens ([b6c233a](https://github.com/Maghwyn/meme-rn-app/commit/b6c233abe311b57f042ca61dbde591c2d0855eee))
+* **env:** get varibles environnements with env.d.ts ([d992988](https://github.com/Maghwyn/meme-rn-app/commit/d9929884280e76b80c79ac14f9fc6a74ec5ed274))
+* **redux-persist:** Add middleware in setup redux store ([dbd1f66](https://github.com/Maghwyn/meme-rn-app/commit/dbd1f668cd3932be166398f9e6c87395d9354edc))
+
+
+### Features
+
+* **auth-ui:** creation de l'ui auth ([dd706a7](https://github.com/Maghwyn/meme-rn-app/commit/dd706a778e76b0c4b507b4dc2329f40b3a8aee79))
+* **auth:** all requests types payload & requests authentification ([5b2823e](https://github.com/Maghwyn/meme-rn-app/commit/5b2823ec50608b6720af6527992e38b587f65085))
+* **axios:** new axios instance client with base url api url ([004d3a9](https://github.com/Maghwyn/meme-rn-app/commit/004d3a96e0de54603c9ecac89a0877f631a9028b))
+* **feed-ui:** creation de l'ui feed ([0595b79](https://github.com/Maghwyn/meme-rn-app/commit/0595b79ac57e354a153b78ca372f35ae75e031c6))
+
+# [1.0.0-dev.2](https://github.com/Maghwyn/meme-rn-app/compare/v1.0.0-dev.1...v1.0.0-dev.2) (2024-03-06)
+
+### Features
+
+-   **babel:** Implement path module resolver ([01081b0](https://github.com/Maghwyn/meme-rn-app/commit/01081b03d1b29296715d973f32fa9306885d506a))
+-   **redux:** Couple asyncstorage and redux to save the data persistently when closing the app ([e1494ff](https://github.com/Maghwyn/meme-rn-app/commit/e1494ff543a2447c636d03cae2775d4ea71c1a57))
+-   **redux:** Initialize redux with the auth and user store + write a comment exampleon how to use redux ([4bab03d](https://github.com/Maghwyn/meme-rn-app/commit/4bab03dd759be1f341e31dec3c38138b289393c5))
+
+# 1.0.0-dev.1 (2024-03-06)
+
+### Bug Fixes
+
+-   **babel:** The babel config prevent ios and android to build because it was missconfigured ([4961996](https://github.com/Maghwyn/meme-rn-app/commit/4961996fe04a73b6740b3402c88f4edb98ace7e5))
+-   **babel:** The babel config prevent ios and android to build because it was missconfigured ([8c3b5c6](https://github.com/Maghwyn/meme-rn-app/commit/8c3b5c6784ead53311e3f73dea37e8dcae990ffa))
+-   **ci:** Ignore the scripts when installing with yarn ([f79e783](https://github.com/Maghwyn/meme-rn-app/commit/f79e783d1f835f3253e7fbc6aa3997b563853dbc))
+-   **prettier:** Prettier failed to work because of missing packages + lint ([99bf4fa](https://github.com/Maghwyn/meme-rn-app/commit/99bf4fa97be7ef6daa966a188cab1faafab30a8a))
+-   **prettier:** Prettier failed to work because of missing packages + lint ([a2fc4bf](https://github.com/Maghwyn/meme-rn-app/commit/a2fc4bf497bcd106b945d58ecf347bf3ceeeaac6))
+-   **yarn:** Update yarn lock file ([31da3ec](https://github.com/Maghwyn/meme-rn-app/commit/31da3ec72a98a39dbb6a851ace9d0ccfc52144eb))
+-   **yarn:** Update yarn lock file ([587f896](https://github.com/Maghwyn/meme-rn-app/commit/587f896caee168437fb66a8eb8682a822bf0edcf))
+-   **yarn:** Update yarn lock file ([4aaf26c](https://github.com/Maghwyn/meme-rn-app/commit/4aaf26c9499a6c0c0acbea5c27a8c9d2b257ac66))
+
+### Features
+
+-   **cd:** Implement the basic cd before deloyment. Wip ci ([9dfd2c4](https://github.com/Maghwyn/meme-rn-app/commit/9dfd2c439a2c21529ea25be81541ec71517dbc99))
+-   **commitling:** Install commitlint for semantic-release support ([5299df9](https://github.com/Maghwyn/meme-rn-app/commit/5299df9e3c038f81736bbcd7edb3b6597e2b53af))
+-   **commitling:** Install commitlint for semantic-release support ([40b164c](https://github.com/Maghwyn/meme-rn-app/commit/40b164c685132f5b909c338d2455df8dfef95d6a))